проверка авт

+++ routes/web.php
@@ -15,25 +15,25 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
     return view('main');
-});
+})->middleware('auth');
 Route::get('map', function () {
     return view('welcome');
-});
+})->middleware('auth');
 
 Route::group(['prefix'=>'admin'],function (){
-    Route::get('users','GKBanAdd@indexuser');
+    Route::get('users','GKBanAdd@indexuser')->middleware('auth');
     Route::get('panel', function () {
         return view('admin.panel');
-    });
-    Route::get('/','GKBanAdd@index');
-    Route::get('/create','GKBanAdd@create');
-    Route::post('/create','GKBanAdd@store');
-    Route::patch('/{id}','GKBanAdd@update');
-    Route::delete('/{id}','GKBanAdd@destroy');
+    })->middleware('auth');
+    Route::get('/','GKBanAdd@index')->middleware('auth');
+    Route::get('/create','GKBanAdd@create')->middleware('auth');
+    Route::post('/create','GKBanAdd@store')->middleware('auth');
+    Route::patch('/{id}','GKBanAdd@update')->middleware('auth');
+    Route::delete('/{id}','GKBanAdd@destroy')->middleware('auth');
 });
 
 Route::group(['prefix'=>'lk'],function(){
-    Route::get('/profile','GKUserProfile@index');
+    Route::get('/profile','GKUserProfile@index')->middleware('auth');
 });
 
 Auth::routes();
